Rename package and compatibility with java 8
