+++ src/wesoga/baseMod/entities/Arrow.java
@@ -1,0 +1,46 @@
+package wesoga.baseMod.entities;
+
+import wesoga.baseMod.BaseMod;
+import wesoga.entities.Entity;
+import wesoga.entities.LivingEntity;
+
+public class Arrow extends Entity {
+	public Arrow(double x, double y, double direction, double throwerSpeedX, double throwerSpeedY) {
+		super(BaseMod.ARROW_MODEL, x, y);
+
+		this.accel(throwerSpeedX - Math.sin(direction) * 0.2, throwerSpeedY - Math.cos(direction) * 0.2);
+	}
+
+	@Override
+	public void onTick() {
+		rotation = Math.atan2(-speedX, -speedY);
+
+		if (speedX * speedX + speedY * speedY < 0.001) {
+			alive = false;
+		}
+	}
+
+	@Override
+	public void onCollision(Entity collidedEntity) {
+		if (alive && collidedEntity instanceof LivingEntity) {
+			((LivingEntity) collidedEntity).damage(2);
+		}
+
+		alive = false;
+	}
+
+	@Override
+	public boolean canWalk() {
+		return false;
+	}
+
+	@Override
+	public boolean canFly() {
+		return true;
+	}
+
+	@Override
+	public double getFriction() {
+		return 0.99;
+	}
+}
