+++ src/wesoga/baseMod/entities/ArrowModel.java
@@ -1,0 +1,10 @@
+package wesoga.baseMod.entities;
+
+import wesoga.entities.EntityModel;
+
+public class ArrowModel extends EntityModel {
+	@Override
+	public double getSize() {
+		return 0.5;
+	}
+}
