+++ src/wesoga/baseMod/entities/Player.java
@@ -1,0 +1,10 @@
+package wesoga.baseMod.entities;
+
+import wesoga.baseMod.BaseMod;
+import wesoga.entities.LivingEntity;
+
+public class Player extends LivingEntity {
+	public Player(double x, double y) {
+		super(BaseMod.PLAYER_MODEL, 10, x, y);
+	}
+}
