+++ src/wesoga/baseMod/entities/PlayerModel.java
@@ -1,0 +1,10 @@
+package wesoga.baseMod.entities;
+
+import wesoga.entities.EntityModel;
+
+public class PlayerModel extends EntityModel {
+	@Override
+	public double getSize() {
+		return 0.9;
+	}
+}
