+++ src/wesoga/baseMod/entities/ZombieModel.java
@@ -1,0 +1,10 @@
+package wesoga.baseMod.entities;
+
+import wesoga.entities.EntityModel;
+
+public class ZombieModel extends EntityModel {
+	@Override
+	public double getSize() {
+		return 0.9;
+	}
+}
