+++ src/wesoga/baseMod/tiles/Grass.java
@@ -1,0 +1,10 @@
+package wesoga.baseMod.tiles;
+
+import wesoga.tiles.TileModel;
+
+public class Grass extends TileModel {
+	@Override
+	public String getID() {
+		return "grass";
+	}
+}
