+++ src/wesoga/entities/EntityModel.java
@@ -1,0 +1,8 @@
+package wesoga.entities;
+
+/**
+ * Store informations needed by the client
+ */
+public abstract class EntityModel {
+	public abstract double getSize();
+}
