--- src/ws/baseMod/entities/ArrowModel.java
@@ -1,10 +1,0 @@
-package ws.baseMod.entities;
-
-import ws.entities.EntityModel;
-
-public class ArrowModel extends EntityModel {
-	@Override
-	public double getSize() {
-		return 0.5;
-	}
-}
