--- src/ws/baseMod/entities/PlayerModel.java
@@ -1,10 +1,0 @@
-package ws.baseMod.entities;
-
-import ws.entities.EntityModel;
-
-public class PlayerModel extends EntityModel {
-	@Override
-	public double getSize() {
-		return 0.9;
-	}
-}
