--- src/ws/baseMod/entities/ZombieModel.java
@@ -1,10 +1,0 @@
-package ws.baseMod.entities;
-
-import ws.entities.EntityModel;
-
-public class ZombieModel extends EntityModel {
-	@Override
-	public double getSize() {
-		return 0.9;
-	}
-}
