--- src/ws/baseMod/tiles/Grass.java
@@ -1,10 +1,0 @@
-package ws.baseMod.tiles;
-
-import ws.tiles.TileModel;
-
-public class Grass extends TileModel {
-	@Override
-	public String getID() {
-		return "grass";
-	}
-}
