--- src/ws/baseMod/tiles/Water.java
@@ -1,16 +1,0 @@
-package ws.baseMod.tiles;
-
-import ws.tiles.TileModel;
-
-public class Water extends TileModel {
-
-	@Override
-	public String getID() {
-		return "water";
-	}
-
-	@Override
-	public boolean isWalkable() {
-		return false;
-	}
-}
