--- src/ws/entities/EntityModel.java
@@ -1,8 +1,0 @@
-package ws.entities;
-
-/**
- * Store informations needed by the client
- */
-public abstract class EntityModel {
-	public abstract double getSize();
-}
